fix: close l7 qa blockers
